Move new InitializeEx functionality to DismApiEx.cs just for organizational purposes
