+++ src/Microsoft.Dism/Microsoft.Dism.csproj
@@ -20,6 +20,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DismApiEx.cs" />
     <Compile Include="DismAppxPackage.cs" />
     <Compile Include="DismCapability.cs" />
     <Compile Include="DismCapabilityInfo.cs" />
